#added repository, renamed packages
