+++ src/main/java/com/yourcodereview/java01/eugenet1024/services/linkGenerator/LinkGeneratorService.java
@@ -1,4 +1,4 @@
-package com.yourcodereview.java01.eugenet1024.service.linkGenerator;
+package com.yourcodereview.java01.eugenet1024.services.linkGenerator;
 
 import com.yourcodereview.java01.eugenet1024.api.linkGenerator.GenerateLinkRequest;
 import com.yourcodereview.java01.eugenet1024.api.linkGenerator.GenerateLinkResponse;
