+++ src/main/java/com/yourcodereview/java01/eugenet1024/services/linkRedirect/LinkRedirectService.java
@@ -1,4 +1,4 @@
-package com.yourcodereview.java01.eugenet1024.service.linkRedirect;
+package com.yourcodereview.java01.eugenet1024.services.linkRedirect;
 
 public interface LinkRedirectService {
