+++ src/main/java/com/yourcodereview/java01/eugenet1024/services/linkRedirect/LinkRedirectServiceImpl.java
@@ -1,4 +1,4 @@
-package com.yourcodereview.java01.eugenet1024.service.linkRedirect;
+package com.yourcodereview.java01.eugenet1024.services.linkRedirect;
 
 import org.springframework.stereotype.Service;
