+++ src/main/java/com/yourcodereview/java01/eugenet1024/services/statistics/StatisticsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.yourcodereview.java01.eugenet1024.service.statistics;
+package com.yourcodereview.java01.eugenet1024.services.statistics;
 
 import com.yourcodereview.java01.eugenet1024.api.statistics.OverallStatisticsDto;
 import com.yourcodereview.java01.eugenet1024.api.statistics.ShortLinkStatisticsDto;
